requirements.txt
aiogram>=3.0.0
Pillow>=10.0.0
PyYAML>=6.0

